+++ src/zendesk_sdk/__init__.py
@@ -1,7 +1,15 @@
 """Interface to Zendesk API."""
 
 from .client import ZendeskServices
-from .exceptions import TicketClosedError
+from .exceptions import TicketClosedError, ZendeskAuthenticationError, ZendeskConfigurationError
 from .models import Attachment, Ticket, TicketComment
 
-__all__ = ["Attachment", "Ticket", "TicketClosedError", "TicketComment", "ZendeskServices"]
+__all__ = [
+    "Attachment",
+    "Ticket",
+    "TicketClosedError",
+    "TicketComment",
+    "ZendeskAuthenticationError",
+    "ZendeskConfigurationError",
+    "ZendeskServices",
+]